+++ isso/js/tests/unit/config.test.js
@@ -9,6 +9,11 @@
 
 "use strict";
 
+beforeEach(() => {
+  jest.resetModules();
+  document.body.innerHTML = '';
+});
+
 test("Client configuration - no languages", () => {
   // Mock navigator.languages = []
   global.languages = jest.spyOn(navigator, "languages", "get")
@@ -34,4 +39,22 @@ test("Client configuration - no languages", () => {
   let expected_langs = ["en", "en"];
 
   expect(config["langs"]).toStrictEqual(expected_langs);
+});
+
+test("data-isso-* i18n strings should be accepted with newline characters", () => {
+
+  document.body.innerHTML =
+      '<div id=isso-thread></div>' +
+      // Note: `src` and `data-isso` need to be set,
+      // else `api` fails to initialize!
+      '<script src="http://isso.api/js/embed.min.js"'
+      + ' data-isso="/"'
+      + '</script>';
+
+  var script_tag = document.getElementsByTagName('script')[0];
+  script_tag.setAttributeNS(null, 'data-isso-num-comments-text-en', "One comment\\n{{ n }} comments");
+
+  const config = require("app/config");
+
+  expect(config['num-comments-text-en']).toMatch("One comment\n{{ n }} comments");
 });