+++ CHANGES.rst
@@ -15,8 +15,6 @@ New Features
 .. _#966: https://github.com/posativ/isso/pull/966
 .. _#998: https://github.com/isso-comments/isso/pull/998
 .. _#1000: https://github.com/isso-comments/isso/pull/1000
-.. _#966: https://github.com/posativ/isso/pull/966
-.. _#998: https://github.com/isso-comments/isso/pull/998
 .. _#1001: https://github.com/isso-comments/isso/pull/1001
 
 Breaking Changes
@@ -34,6 +32,7 @@ Bugfixes & Improvements
 - Fix W3C Validation issues (`#999`_, pkvach)
 - Handle deleted comments in Disqus migration (`#994`_, pkvach)
 - Fix total comments count calculation (`#997`_, pkvach)
+- Fix newline character handling in data-isso-* i18n strings (`#992`_, pkvach)
 
 .. _#951: https://github.com/posativ/isso/pull/951
 .. _#967: https://github.com/posativ/isso/pull/967
@@ -42,6 +41,7 @@ Bugfixes & Improvements
 .. _#999: https://github.com/isso-comments/isso/pull/999
 .. _#994: https://github.com/isso-comments/isso/pull/994
 .. _#997: https://github.com/isso-comments/isso/pull/997
+.. _#992: https://github.com/isso-comments/isso/pull/992
 
 0.13.1.dev0 (2023-02-05)
 ------------------------